+++ src/commands/buffer.c
@@ -7,20 +7,14 @@
 #include <parser/parser.h>
 #include <errors/errors.h>
 
-void	init_buffer(t_buffer *buffer)
-{
-	ft_bzero(&buffer->buf[0], BUFFER_SIZE);
-	buffer->index = 0;
-}
-
-void	append_char_to_buffer(const char **start, t_buffer *buffer)
+static void	append_char_to_buffer(const char **start, t_buffer *buffer)
 {
 	buffer->buf[buffer->index] = **start;
 	++(buffer->index);
 	++(*start);
 }
 
-void	append_env_value_to_buffer(const char **start, \
+static void	append_env_value_to_buffer(const char **start, \
 									t_buffer *buffer, \
 									t_bool should_trim)
 {
@@ -51,7 +45,7 @@ void	append_env_value_to_buffer(const char **start, \
 	*start = *start + key_len;
 }
 
-void	append_quoted_string_to_buffer(const char **start, t_buffer *buffer)
+static void	append_quoted_string_to_buffer(const char **start, t_buffer *buffer)
 {
 	const t_quotes_position	quotes = get_quotes_positions(*start);
 
@@ -71,15 +65,10 @@ void	append_quoted_string_to_buffer(const char **start, t_buffer *buffer)
 		++(*start);
 }
 
-void	append_exit_code_to_buffer(const char **start, t_buffer *buffer)
+void	init_buffer(t_buffer *buffer)
 {
-	const char *exit_code_string = ft_itoa((int)*get_return_code());
-	const char *ptr = exit_code_string;
-
-	while (*exit_code_string)
-		append_char_to_buffer(&exit_code_string, buffer);
-	*start += 2;
-	free((void *)ptr);
+	ft_bzero(&buffer->buf[0], BUFFER_SIZE);
+	buffer->index = 0;
 }
 
 void	append_expanded_input_to_buffer(t_arg *arg, t_buffer *buffer)
