+++ src/errors/errors.c
@@ -7,22 +7,6 @@
 #include <output/write_to_std.h>
 #include <executor/run_commands.h>
 
-t_exit_code	*get_return_code(void)
-{
-	static t_exit_code	code = 0;
-
-	return (&code);
-}
-
-t_exit_code	set_return_code(t_exit_code new_code)
-{
-	t_exit_code	*code;
-
-	code = get_return_code();
-	*code = new_code;
-	return (*code);
-}
-
 static t_bool	should_exit(t_exit_code code)
 {
 	if (code == MALLOC_ERROR || \
@@ -30,7 +14,7 @@ static t_bool	should_exit(t_exit_code code)
 		code == FORK_ERROR || \
 		code == PIPE_ERROR)
 	{
-		set_return_code(SYS_ERROR);
+		set_exit_code(SYS_ERROR);
 		return (TRUE);
 	}
 	return (FALSE);
@@ -40,8 +24,7 @@ static void	write_system_error(t_exit_code code, \
 								const char *location, \
 								const char *filename)
 {
-	if (code == FILE_ERROR || \
-		code == SYS_ERROR || \
+	if (code == FILE_ERROR || code == SYS_ERROR || \
 		(code >= MALLOC_ERROR && code <= PIPE_ERROR))
 	{
 		if (location)
@@ -56,7 +39,7 @@ static void	write_system_error(t_exit_code code, \
 		}
 		write_to_stderr(strerror(errno));
 		write_to_stderr("\n");
-		set_return_code(SYS_ERROR);
+		set_exit_code(SYS_ERROR);
 	}
 }
 
@@ -66,7 +49,7 @@ void	handle_error(t_exit_code code, \
 {
 	if (code == SUCCESS)
 		return ;
-	set_return_code(code);
+	set_exit_code(code);
 	write_to_stderr("BestShellEver: ");
 	if (code == HOME_NOT_SET_ERROR)
 	{
@@ -81,5 +64,5 @@ void	handle_error(t_exit_code code, \
 	else
 		write_system_error(code, location, filename);
 	if (should_exit(code))
-		exit(*get_return_code());
+		exit(*get_exit_code());
 }
